requirements.txt
setuptools~=47.3.1
python-vlc
pynput~=1.6.8
eyeD3~=0.9.5